Corrected handling or ERROR state, enforced setting scio to 0 in IDLE
The former implementation in case of error remained in the S_ERROR state until reset.
In certain scenarios that may be useful. However, in most cases it is better if the controller returns to S_IDLE with error flag set.
I have also added setting scio to 0 when controller remains in S_IDLE state.